Remove awaitility and use the TestContext.
Especially since RunTestOnContext is used, blocking the test method is
not a great idea. While the tests were not wrong per se, I did end up
blocking the test method when writing the auto-dispatch code, which
makes me suspicious that a lot of the existing tests were assuming/
taking advantage of the fact that the idLoader/et all were immediately
resolved promises.

I might try and make idLoader use a runOnContext to make sure the
tests are actually testing "some happens later", but I'll defer
this to a follow up commit.